+++ CHANGELOG.md
@@ -41,6 +41,7 @@
 - Added a header file defining common math functions (e.g., sigmoid) to be used throughout the code.
 - Added capability to print monitored variables in multiple formats, triggered from `Ida::runSimulation`.
 - Added IDA statistics object which can be accumulated over multiple simulations.
+- Minor performance improvements to residual evaluation in PowerElectronics module.
 
 ## v0.1